try to share imagecdn tests
