+++ e2e-tests/development-runtime/gatsby-config.js
@@ -72,6 +72,7 @@ module.exports = {
     `gatsby-plugin-sass`,
     `gatsby-plugin-less`,
     `gatsby-plugin-stylus`,
+    `gatsby-cypress/remote-file/plugin`,
   ],
   partytownProxiedURLs: [
     `http://localhost:8888/three.js`,
