+++ e2e-tests/production-runtime/gatsby-config.js
@@ -32,6 +32,7 @@ module.exports = {
     `gatsby-plugin-sass`,
     `gatsby-plugin-less`,
     `gatsby-plugin-stylus`,
+    `gatsby-cypress/remote-file/plugin`,
   ].concat(process.env.TEST_PLUGIN_OFFLINE ? [`gatsby-plugin-offline`] : []),
   partytownProxiedURLs: [
     `http://localhost:8888/three.js`,
