+++ packages/gatsby-cypress/remote-file/plugin/gatsby-config.js
@@ -1,0 +1,9 @@
+/** @type {import("gatsby").GatsbyConfig} */
+const config = {
+  plugins: [
+    `gatsby-plugin-image`,
+    `gatsby-plugin-sharp`,
+  ]
+}
+
+module.exports = config
