+++ packages/gatsby-cypress/remote-file/plugin/package.json
@@ -1,0 +1,3 @@
+{
+  "name": "gatsby-plugin-e2e-remote-file"
+}
